Remove TextTagIssueSection.
